requirements.txt
fastapi
uvicorn
praw
openai
python-dotenv
markdown
pdfkit
